CFY-6999 Add ARIA AWS hello world example
The example includes an instance that is connected to an elastic ip,
a security group, and a keypair. The instance hosts a web application
node that sets up a server which displays the ARIA logo at a designated port. The
port is provided using the TOSCA input mechanism.

This example forced some small changes in the context adapter, as
we had to rely on the Cloudify AWS plugin code, which sometimes
requires very specific YAML entity names in order to operate correctly.