CFY-7056 Repository refactoring
The cloudify-aria-extensions repository is meant to help ARIA to utilize
some resources written by the Cloudify team, such as AWS and Openstack
integration support (using the Cloudify AWS and Openstack plugins).

As a result, some older components of this repository, such as ARIA
parsing abilities of the Cloudify DSL, are no longer relevant to this
repository. In addition, additional refactoring was needed in order to better
represent the purpose of this repository.